Require_activation now defaults to false, update and clarify README (#534)
Previously, require_activation was set to true by default, resulting in
this warning message:

RequireActivation is set to true, but the following required settings are missing: EmailFrom, EmailHost, HostURL

Default behavior producing warning or error messages is confusing and
should not happen. The only benefit of defaulting to true is if the
error message helps point to the additinal settings. But a better way to
inform the user is to clarify it in the README.