favicon: crate favicon file only after retreival has succeded
Previously, an empty file was created at the start of the downloadIcon
function and would remain as an empty file even when favicon retreival
failed. This empty file would prevent future attempts to download an
icon for the same site ID.

Now the icon file is created at the end, after everything else has
succeeded.