More reliable favicon retrieval. (#652)
* More reliable favicon retrieval.

Previously, the application looked for favicons in `/favicon.ico`.
But favicons can be anywhere on a website and the location can be
defined in HTML and JSON manifests. Websites can also have multiple
favicons.

As a result, when a new site was added to stash-box which did not have a
`/favicon.ico`, the site was created without a proper favicon. Additionally,
since stash-box didn't check the filetype of the data returned by the
`/favicon.ico` GET request, it stored various 404 and other redirected
html pages.

A non-trivial amount of logic is required to discover all favicon
locations, check filetypes, and sort them. It therefore makes sense to
rely on a third-party package for this job. This patch uses
`go.deanishe.net/favicon`.

* Check length of icons to avoid runtime error when using icons[0] later.

* Specify go.deanishe.net/favicon as direct requirement

* favicon: crate favicon file only after retreival has succeded

Previously, an empty file was created at the start of the downloadIcon
function and would remain as an empty file even when favicon retreival
failed. This empty file would prevent future attempts to download an
icon for the same site ID.

Now the icon file is created at the end, after everything else has
succeeded.

* retreive the smallest available favicon for sites

* use http.client with cookiejar for favicon finder

Cookies are required for certain websites, even for simple requests.
See code comment for details.

* Makefile: include generate in the default make target

This more closely mirrors the GitHub workflow build process and helps
avoid CI build failures on pull requests.